站点签到移植为插件
