+++ app/plugins/modules/sites/chdbits.py
@@ -6,7 +6,7 @@ import re
 from lxml import etree
 
 from app.helper.openai_helper import OpenAiHelper
-from app.sites.sitesignin._base import _ISiteSigninHandler
+from app.plugins.modules.sites._base import _ISiteSigninHandler
 from app.utils import StringUtils, RequestUtils
 from config import Config
