+++ app/plugins/modules/sites/hdarea.py
@@ -1,19 +1,19 @@
-import json
-
-from app.sites.sitesignin._base import _ISiteSigninHandler
+from app.plugins.modules.sites._base import _ISiteSigninHandler
 from app.utils import StringUtils, RequestUtils
 from config import Config
 
 
-class PTerClub(_ISiteSigninHandler):
+class HDArea(_ISiteSigninHandler):
     """
-    猫签到
+    好大签到
     """
+    
     # 匹配的站点Url，每一个实现类都需要设置为自己的站点Url
-    site_url = "pterclub.com"
+    site_url = "hdarea.co"
 
-    # 已签到
-    _sign_regex = ['签到已得\\d+']
+    # 签到成功
+    _success_text = "此次签到您获得"
+    _repeat_text = "请不要重复签到哦"
 
     @classmethod
     def match(cls, url):
@@ -35,23 +35,29 @@ class PTerClub(_ISiteSigninHandler):
         ua = site_info.get("ua")
         proxy = Config().get_proxies() if site_info.get("proxy") else None
 
-        # 签到
-        sign_res = RequestUtils(cookies=site_cookie,
+        # 获取页面html
+        data = {
+            'action': 'sign_in'
+        }
+        html_res = RequestUtils(cookies=site_cookie,
                                 headers=ua,
                                 proxies=proxy
-                                ).get_res(url="https://pterclub.com/attendance-ajax.php")
-        if not sign_res or sign_res.status_code != 200:
-            self.error(f"签到失败，签到接口请求失败")
-            return False, f'【{site}】签到失败，请检查cookie是否失效'
+                                ).post_res(url="https://www.hdarea.co/sign_in.php", data=data)
+        if not html_res or html_res.status_code != 200:
+            self.error(f"签到失败，请检查站点连通性")
+            return False, f'【{site}】签到失败，请检查站点连通性'
 
-        sign_dict = json.loads(sign_res.text)
-        self.debug(f"签到返回 {sign_dict}")
-        if sign_dict['status'] == 1:
-            # {"status":"1","data":" (签到已成功300)","message":"<p>这是您的第<b>237</b>次签到，
-            # 已连续签到<b>237</b>天。</p><p>本次签到获得<b>300</b>克猫粮。</p>"}
+        if "login.php" in html_res.text:
+            self.error(f"签到失败，cookie失效")
+            return False, f'【{site}】签到失败，cookie失效'
+
+        # 判断是否已签到
+        # '已连续签到278天，此次签到您获得了100魔力值奖励!'
+        if self._success_text in html_res.text:
             self.info(f"签到成功")
             return True, f'【{site}】签到成功'
-        else:
-            # {"status":"0","data":"抱歉","message":"您今天已经签到过了，请勿重复刷新。"}
+        if self._repeat_text in html_res.text:
             self.info(f"今日已签到")
             return True, f'【{site}】今日已签到'
+        self.error(f"签到失败，签到接口返回 {html_res.text}")
+        return False, f'【{site}】签到失败'
