+++ app/plugins/modules/sites/hdcity.py
@@ -1,17 +1,19 @@
-from app.sites.sitesignin._base import _ISiteSigninHandler
+from app.plugins.modules.sites._base import _ISiteSigninHandler
 from app.utils import StringUtils, RequestUtils
 from config import Config
 
 
-class HaiDan(_ISiteSigninHandler):
+class HDCity(_ISiteSigninHandler):
     """
-    海胆签到
+    城市签到
     """
     # 匹配的站点Url，每一个实现类都需要设置为自己的站点Url
-    site_url = "haidan.video"
+    site_url = "hdcity.city"
 
     # 签到成功
-    _success_text = "已经打卡"
+    _success_text = '本次签到获得魅力'
+    # 重复签到
+    _repeat_text = '已签到'
 
     @classmethod
     def match(cls, url):
@@ -33,22 +35,16 @@ class HaiDan(_ISiteSigninHandler):
         ua = site_info.get("ua")
         proxy = Config().get_proxies() if site_info.get("proxy") else None
 
-        # 签到
-        RequestUtils(cookies=site_cookie,
-                     headers=ua,
-                     proxies=proxy
-                     ).get(url="https://www.haidan.video/signin.php")
-
         # 获取页面html
         html_res = RequestUtils(cookies=site_cookie,
                                 headers=ua,
                                 proxies=proxy
-                                ).get_res(url="https://www.haidan.video")
+                                ).get_res(url="https://hdcity.city/sign")
         if not html_res or html_res.status_code != 200:
             self.error(f"签到失败，请检查站点连通性")
             return False, f'【{site}】签到失败，请检查站点连通性'
 
-        if "login.php" in html_res.text:
+        if "login" in html_res.text:
             self.error(f"签到失败，cookie失效")
             return False, f'【{site}】签到失败，cookie失效'
 
@@ -57,6 +53,8 @@ class HaiDan(_ISiteSigninHandler):
         if self._success_text in html_res.text:
             self.info(f"签到成功")
             return True, f'【{site}】签到成功'
-
+        if self._repeat_text in html_res.text:
+            self.info(f"今日已签到")
+            return True, f'【{site}】今日已签到'
         self.error(f"签到失败，签到接口返回 {html_res.text}")
         return False, f'【{site}】签到失败'
