+++ app/plugins/modules/sites/hdsky.py
@@ -2,7 +2,7 @@ import json
 import time
 
 from app.helper import OcrHelper
-from app.sites.sitesignin._base import _ISiteSigninHandler
+from app.plugins.modules.sites._base import _ISiteSigninHandler
 from app.utils import StringUtils, RequestUtils
 from config import Config
