+++ app/sites/__init__.py
@@ -1,7 +1,6 @@
 from app.sites.site_userinfo import SiteUserInfo
 from .sites import Sites
 from .site_cookie import SiteCookie
-from .site_signin import SiteSignin
 from .site_subtitle import SiteSubtitle
 from .siteconf import SiteConf
 from .site_limiter import SiteRateLimiter
