+++ web/templates/service.html
@@ -276,42 +276,7 @@
     </div>
   </div>
 </div>
-<div class="modal modal-blur fade" id="modal-service-signin" tabindex="-1" role="dialog" aria-hidden="true"
-     data-bs-backdrop="static" data-bs-keyboard="false">
-  <div class="modal-dialog modal-lg modal-dialog-centered" role="document">
-    <div class="modal-content">
-      <div class="modal-header">
-        <h5 class="modal-title">手动站点签到</h5>
-        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-      </div>
-      <div class="modal-body">
-        <div class="row">
-          <div class="col">
-            <div class="mb-3">
-              <div class="btn-list">
-                  <label class="form-label">站点 <span class="form-help"
-                                                        title="不选择时默认全部运行"
-                                                        data-bs-toggle="tooltip">?</span></label>
-                  <a href="javascript:void(0)" class="ms-auto" onclick="select_btn_SelectALL(this, 'service_site_signin')">全选</a>
-              </div>
-              <div class="form-selectgroup">
-                {% for SiteInfo in SigninSites %}
-                  <label class="form-selectgroup-item">
-                    <input type="checkbox" name="service_site_signin" value="{{ SiteInfo.id }}" class="form-selectgroup-input">
-                    <span class="form-selectgroup-label">{{ SiteInfo.name }}</span>
-                  </label>
-                {% endfor %}
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div class="modal-footer">
-        <a href="javascript:run_signin_now()" id="service_signin_btn" class="btn btn-primary">开始签到</a>
-      </div>
-    </div>
-  </div>
-</div>
+
 <script type="text/javascript">
 
   // 显示服务提示框
@@ -352,9 +317,6 @@
         break;
       case "sync":
         $('#modal-service-sync').modal('show');
-        break;
-      case "ptsignin":
-        $('#modal-service-signin').modal('show');
         break;
       default:
         show_ask_modal("是否立即运行 " + name + "？", function () {
@@ -547,20 +509,8 @@
     });
   }
 
-  // 站点签到
-  function run_signin_now() {
-    let sids = select_GetSelectedVAL("service_site_signin");
-    ajax_post("run_signin", {sids: sids}, function (ret) {
-      $("#modal-service-signin").modal('hide');
-      if (ret.code === 0) {
-        show_success_modal("站点签到已启动，正在后台运行");
-      } else {
-        show_fail_modal(ret.msg);
-      }
-    });
-  }
-
   // 初始化DropZone
   backup_dropzone = new Dropzone("#back_file");
   backup_dropzone.options.acceptedFiles = ".zip";
+
 </script>
