+++ web/templates/site/site.html
@@ -84,9 +84,6 @@
           </dl>
           <div class="row">
             <div class="col">
-              {% if Site.signin_enable and (Site.signurl or Site.rssurl) %}
-              <span class="badge bg-blue-lt mb-1 me-1">签到</span>
-              {% endif %}
               {% if Site.rss_enable and Site.rssurl %}
               <span class="badge bg-azure-lt mb-1 me-1">订阅</span>
               {% endif %}
@@ -156,7 +153,7 @@
           <div class="col-lg-6">
             <div class="mb-3">
               <label class="form-label required">站点地址 <span class="form-help"
-                  title="站点访问地址，自动签到、站点数据统计、刷流、内置索引器等将使用，有独立签到页面的站点可以配置为签到页面以实现自动签到，否则可以配置为首页（配置为首页时可通过开启站点浏览器仿真来实现自动签到）"
+                  title="站点访问地址，站点签到、数据统计、刷流、内置索引器等将使用"
                   data-bs-toggle="tooltip">?</span></label>
               <input type="text" class="form-control" id="site_signurl" value="" placeholder="站点http地址">
             </div>
@@ -166,10 +163,6 @@
               <label class="form-label required">站点用途 <span class="form-help" title="设置站点使用场景，选中的场景才会使用该站点"
                   data-bs-toggle="tooltip">?</span></label>
               <div class="form-selectgroup">
-                <label class="form-selectgroup-item">
-                  <input type="checkbox" name="site_uses" value="Q" class="form-selectgroup-input" checked>
-                  <span class="form-selectgroup-label">签到</span>
-                </label>
                 <label class="form-selectgroup-item">
                   <input type="checkbox" name="site_uses" value="D" class="form-selectgroup-input" checked>
                   <span class="form-selectgroup-label">订阅</span>
